+++ CHANGES.md
@@ -33,6 +33,7 @@
 * Change restart to use loading page
 * Add upgrade messages for sickbeard, cache, and failed db upgrade processes to loading page
 * Change add WorkingDirectory to systemd startup prevents startup git issue
+* Change improve MagnetDLProvider latest releases search
 
 
 [develop changelog]